// cypress/support/pageObjects/dataEntryForm/enums/sectionFormEnum.ts
export enum SectionFormTitles {
    SECTION_ONE_TITLE = "Section 1",
    SECTION_TWO_TITLE = "Section 2",
    SECTION_TREE_TITLE = "Section 3"
}

export enum SectionFormDescription {
    SECTION_ONE_DESCRIPTION = "This is a paragraph in Section 1.",
    SECTION_TWO_DESCRIPTION = "This is a paragraph in Section 2.",
    SECTION_TREE_DESCRIPTION = "This is a paragraph in Section 3."
}