// cypress/support/pageObjects/dataEntryForm/interfaces/sectionFormInterface.ts
export interface SectionFormLocatorsInterface {
    sectionOneTextLocator: string;
    paraOneTextLocator: string;
    inputOneTextLocator: string;
    submitButtonOneLocator: string;
}

export interface SectionFormSectionsInterface {
    sectionOne: SectionFormLocatorsInterface
    sectionTwo: SectionFormLocatorsInterface
    sectionTree: SectionFormLocatorsInterface
}